+++ src/main/java/hello/cluebackend/application/notice/dto/response/NoticeInfoDto.java
@@ -1,9 +1,12 @@
-package hello.cluebackend.domain.notice.controller.dto.response;
+package hello.cluebackend.application.notice.dto.response;
 
 import hello.cluebackend.domain.notice.model.NoticeType;
+import hello.cluebackend.domain.noticedocument.controller.dto.response.NoticeDocumentDto;
 import lombok.*;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.UUID;
 
 @Getter
@@ -11,10 +14,11 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class NoticeDto {
+public class NoticeInfoDto {
     private NoticeType type;
     private UUID noticeId;
     private String title;
     private String content;
     private LocalDateTime createdAt;
+    List<NoticeDocumentDto> noticeDocuments = new ArrayList<>();
 }
