+++ src/main/java/hello/cluebackend/domain/notice/service/NoticeCommandService.java
@@ -1,36 +1,29 @@
-
 package hello.cluebackend.domain.notice.service;
 
 import hello.cluebackend.application.document.dto.UrlDto;
+import hello.cluebackend.application.notice.dto.request.AddNoticeDto;
+import hello.cluebackend.application.notice.dto.request.CreateNoticeDto;
+import hello.cluebackend.application.notice.dto.request.ModifyNoticeDto;
+import hello.cluebackend.application.notice.dto.request.NoticeFileDto;
 import hello.cluebackend.domain.file.service.FileService;
-import hello.cluebackend.domain.notice.model.Notice;
-import hello.cluebackend.domain.notice.model.repository.NoticeRepository;
-import hello.cluebackend.domain.notice.controller.dto.request.AddNoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.request.CreateNoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.request.ModifyNoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.request.NoticeFileDto;
-import hello.cluebackend.domain.notice.controller.dto.response.NoticeDto;
-import hello.cluebackend.domain.notice.controller.dto.response.NoticeInfoDto;
 import hello.cluebackend.domain.notice.exception.IsNotMyNoticeException;
+import hello.cluebackend.domain.notice.model.Notice;
 import hello.cluebackend.domain.noticedocument.domain.FileType;
 import hello.cluebackend.domain.noticedocument.domain.NoticeDocument;
 import hello.cluebackend.domain.noticedocument.domain.repository.NoticeDocumentRepository;
-import hello.cluebackend.domain.noticedocument.controller.dto.response.NoticeDownloadDto;
-import hello.cluebackend.domain.noticedocument.controller.dto.response.NoticeUrlDto;
 import hello.cluebackend.domain.user.model.UserEntity;
 import hello.cluebackend.infrastructure.persistence.classroom.ClassRoomJpaRepository;
+import hello.cluebackend.infrastructure.persistence.notice.NoticeJpaRepository;
 import hello.cluebackend.infrastructure.persistence.user.UserJpaRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -39,11 +32,10 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class NoticeService {
+public class NoticeCommandService {
 
-    private final NoticeRepository noticeRepository;
+    private final NoticeJpaRepository noticeJpaRepository;
     private final NoticeDocumentRepository noticeDocumentRepository;
-    private final ClassRoomJpaRepository classRoomRepository;
     private final UserJpaRepository userRepository;
     private final FileService fileService;
 
@@ -60,7 +52,7 @@ public class NoticeService {
                 .type(dto.getType())
                 .build();
 
-        noticeRepository.save(notice);
+        noticeJpaRepository.save(notice);
 
         em.flush();
         em.clear();
@@ -110,57 +102,13 @@ public class NoticeService {
         if(!isMyNotice(userId, noticeId)) {
             throw new IsNotMyNoticeException("해당 공지사항을 수정할 권한이 없습니다.");
         }
-        Notice notice = noticeRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
+        Notice notice = noticeJpaRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
         List<NoticeDocument> noticeDocuments = noticeDocumentRepository.findByNotice_NoticeId(noticeId);
         for(NoticeDocument noticeDocument : noticeDocuments){
             fileService.deleteFile(noticeDocument.getValue());
             noticeDocumentRepository.delete(noticeDocument);
         }
-        noticeRepository.delete(notice);
-    }
-
-    @Transactional(readOnly = true)
-    public List<NoticeDto> findAllById() {
-        return noticeRepository.findAll().stream()
-                .map(Notice::toDto).toList();
-    }
-
-    @Transactional(readOnly = true)
-    public NoticeInfoDto findById(UUID userId, UUID noticeId) {
-        if(!isMyNotice(userId, noticeId)) {
-            throw new IsNotMyNoticeException("해당 공지사항을 수정할 권한이 없습니다.");
-        }
-        Notice notice = noticeRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
-        List<NoticeDocument> noticeDocument = notice.getNoticeDocuments();
-        NoticeInfoDto noticeInfoDto = notice.toInfoDto();
-        noticeInfoDto.setNoticeDocuments(noticeDocument.stream().map(NoticeDocument::toDto).toList());
-        return noticeInfoDto;
-    }
-
-    @Transactional(readOnly = true)
-    public NoticeDownloadDto downloadNoticeDocument(UUID noticeDocumentId) throws IOException {
-        NoticeDocument noticeDocument = noticeDocumentRepository.findById(noticeDocumentId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항 파일을 찾을 수가 없습니다."));
-        if(noticeDocument.getType() == FileType.FILE){
-            Resource resource = fileService.downloadFile(noticeDocument.getValue());
-            return NoticeDownloadDto.builder()
-                    .original(noticeDocument.getValue())
-                    .contentType(noticeDocument.getContentType())
-                    .resource(resource)
-                    .build();
-        }
-        else return null;
-    }
-
-    @Transactional(readOnly = true)
-    public NoticeUrlDto getLink(UUID noticeDocumentId) {
-        NoticeDocument noticeDocument = noticeDocumentRepository.findById(noticeDocumentId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항 링크를 찾을 수가 없습니다."));
-        if(noticeDocument.getType() == FileType.URL) {
-            return NoticeUrlDto.builder()
-                    .value(noticeDocument.getValue())
-                    .title(noticeDocument.getTitle())
-                    .build();
-        }
-        return null;
+        noticeJpaRepository.delete(notice);
     }
 
     public void removeNoticeDocument(UUID userId, UUID noticeId, UUID noticeDocumentId) {
@@ -178,7 +126,7 @@ public class NoticeService {
         if(!isMyNotice(userId, noticeId)) {
             throw new IsNotMyNoticeException("해당 공지사항을 수정할 권한이 없습니다.");
         }
-        Notice notice = noticeRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
+        Notice notice = noticeJpaRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
         notice.modify(modifyNoticeDto);
     }
 
@@ -186,7 +134,7 @@ public class NoticeService {
         if(!isMyNotice(userId, noticeId)) {
             throw new IsNotMyNoticeException("해당 공지사항을 수정할 권한이 없습니다.");
         }
-        Notice notice = noticeRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
+        Notice notice = noticeJpaRepository.findById(noticeId).orElseThrow(() -> new EntityNotFoundException("해당 공지사항이 찾을 수가 없습니다."));
         if(dto.getFileInfo().size() != files.size()){
             throw new RuntimeException("한쪽 요소 부족");
         }
@@ -221,7 +169,7 @@ public class NoticeService {
     }
 
     public boolean isMyNotice(UUID userId, UUID noticeId) {
-        boolean isMyNotice = noticeRepository.findMyNoticeByUserId(userId, noticeId) > 0;
+        boolean isMyNotice = noticeJpaRepository.findMyNoticeByUserId(userId, noticeId) > 0;
         log.info("isMyNotice: {}", isMyNotice);
         return isMyNotice;
     }
