+++ src/main/java/hello/cluebackend/infrastructure/persistence/notice/NoticeJpaRepository.java
@@ -1,4 +1,4 @@
-package hello.cluebackend.domain.notice.model.repository;
+package hello.cluebackend.infrastructure.persistence.notice;
 
 import hello.cluebackend.domain.notice.model.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.UUID;
 
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, UUID> {
+public interface NoticeJpaRepository extends JpaRepository<Notice, UUID> {
     @Query("select n from Notice n" +
             " join fetch n.user u" +
             " where u.userId = :userId")
